// src/api/generated/model/productCreateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 뭐든배달 API 명세서
 * 뭐든배달 API 명세서입니다
 * OpenAPI spec version: v1
 */

export interface ProductCreateRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  /**
   * @minLength 0
   * @maxLength 500
   */
  description?: string;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  price: number;
  /**
   * @minLength 1
   * @maxLength 256
   */
  imageUrl: string;
  /** @minimum 0 */
  initialStock: number;
}